% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bauw-package.R
\docType{package}
\name{bauw-package}
\alias{bauw}
\alias{bauw-package}
\title{bauw: Data from Bayesian Analysis using Winbugs}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/bauw}
  \item \url{https://poissonconsulting.github.io/bauw/}
  \item Report bugs at \url{https://github.com/poissonconsulting/bauw/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

}
\keyword{internal}
