% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hm.R
\docType{data}
\name{hm}
\alias{hm}
\title{House martin annual counts}
\format{A data frame with 20 rows and 2 columns}
\usage{
hm
}
\description{
The house martin (\emph{Delichon urbica}) population
annual counts from Magden (a small village in Northern Switzerland)
collected by Reto Freuler from 1990 to 2009.
}
\details{
The variables are as follows:
\itemize{
  \item \code{hm} the count (integer).
  \item \code{year} the year (integer).
}
}
\references{
Kery M & Schaub M (2011) Bayesian Population Analysis
using WinBUGS. Academic Press. (\url{http://www.vogelwarte.ch/bpa})
}
\keyword{datasets}
